+++ quantify/data/handling.py
@@ -3,20 +3,22 @@
 # Repository:     https://gitlab.com/quantify-os/quantify-core
 # Copyright (C) Qblox BV & Orange Quantum Systems Holding BV (2020-2021)
 # -----------------------------------------------------------------------------
+from __future__ import annotations
 import os
 import sys
 import json
+from typing import Union
+from collections import OrderedDict
 from collections.abc import Iterable
-from datetime import datetime
+import datetime
 from uuid import uuid4
+from dateutil.parser import parse
+
 import numpy as np
 import xarray as xr
 from qcodes import Instrument
 from quantify.data.types import TUID
-from quantify.utilities.general import (
-    delete_keys_from_dict,
-    get_keys_containing,
-)
+from quantify.utilities.general import delete_keys_from_dict
 
 # this is a pointer to the module object instance itself.
 this = sys.modules[__name__]
@@ -24,13 +26,13 @@ this = sys.modules[__name__]
 this._datadir = None
 
 
-def gen_tuid(ts=None):
+def gen_tuid(ts: datetime.datetime = None) -> TUID:
     """
     Generates a :class:`~quantify.data.types.TUID` based on current time.
 
     Parameters
     ----------
-    ts : :class:`datetime.datetime`
+    ts: :class:`~datetime.datetime`
         optional, can be passed to ensure the tuid is based on a specific time.
 
     Returns
@@ -39,13 +41,13 @@ def gen_tuid(ts=None):
         timestamp based uid.
     """
     if ts is None:
-        ts = datetime.now()
+        ts = datetime.datetime.now()
     # ts gives microsecs by default
-    (dt, micro) = ts.strftime("%Y%m%d-%H%M%S-.%f").split(".")
+    (date_time, micro) = ts.strftime("%Y%m%d-%H%M%S-.%f").split(".")
     # this ensures the string is formatted correctly as some systems return 0 for micro
-    dt = "%s%03d-" % (dt, int(micro) / 1000)
+    date_time = f"{date_time}{int(int(micro) / 1000):03d}-"
     # the tuid is composed of the timestamp and a 6 character uuid.
-    tuid = TUID(dt + str(uuid4())[:6])
+    tuid = TUID(date_time + str(uuid4())[:6])
 
     return tuid
 
@@ -76,7 +78,7 @@ def get_datadir():
     return this._datadir
 
 
-def set_datadir(datadir: str):
+def set_datadir(datadir: str) -> None:
     """
     Sets the data directory.
 
@@ -100,7 +102,7 @@ def _locate_experiment_file(tuid: TUID, datadir: str, name: str) -> str:
     exp_folders = list(filter(lambda x: tuid in x, os.listdir(daydir)))
     if len(exp_folders) == 0:
         print(os.listdir(daydir))
-        raise FileNotFoundError("File with tuid: {} was not found.".format(tuid))
+        raise FileNotFoundError(f"File with tuid: {tuid} was not found.")
 
     # We assume that the length is 1 as tuid is assumed to be unique
     exp_folder = exp_folders[0]
@@ -196,7 +198,9 @@ def create_exp_folder(tuid: TUID, name: str = "", datadir=None):
     return exp_folder
 
 
-def initialize_dataset(settable_pars, setpoints, gettable_pars):
+def initialize_dataset(
+    settable_pars: Iterable, setpoints: np.ndarray, gettable_pars: Iterable
+):
     """
     Initialize an empty dataset based on settable_pars, setpoints and gettable_pars
 
@@ -213,19 +217,20 @@ def initialize_dataset(settable_pars, setpoints, gettable_pars):
     :class:`xarray.Dataset`
         the dataset
     """
+
     darrs = []
+    coords = []
     for i, setpar in enumerate(settable_pars):
-        darrs.append(
-            xr.DataArray(
-                data=setpoints[:, i],
-                name="x{}".format(i),
-                attrs={
-                    "name": setpar.name,
-                    "long_name": setpar.label,
-                    "units": setpar.unit,
-                },
-            )
-        )
+        attrs = {
+            "name": setpar.name,
+            "long_name": setpar.label,
+            "units": setpar.unit,
+            "batched": _is_batched(setpar),
+        }
+        if attrs["batched"] and hasattr(setpar, "batch_size"):
+            attrs["batch_size"] = getattr(setpar, "batch_size")
+        coords.append(f"x{i}")
+        darrs.append(xr.DataArray(data=setpoints[:, i], name=coords[-1], attrs=attrs))
 
     numpoints = len(setpoints[:, 0])
     j = 0
@@ -239,24 +244,33 @@ def initialize_dataset(settable_pars, setpoints, gettable_pars):
 
         count = 0
         for idx, info in enumerate(itrbl):
+            attrs = {
+                "name": info[0],
+                "long_name": info[1],
+                "units": info[2],
+                "batched": _is_batched(getpar),
+            }
+            if attrs["batched"] and hasattr(getpar, "batch_size"):
+                attrs["batch_size"] = getattr(getpar, "batch_size")
             empty_arr = np.empty(numpoints)
             empty_arr[:] = np.nan
             darrs.append(
                 xr.DataArray(
                     data=empty_arr,
-                    name="y{}".format(j + idx),
-                    attrs={"name": info[0], "long_name": info[1], "units": info[2]},
+                    name=f"y{j + idx}",
+                    attrs=attrs,
                 )
             )
             count += 1
         j += count
 
     dataset = xr.merge(darrs)
+    dataset = dataset.set_coords(coords)
     dataset.attrs["tuid"] = gen_tuid()
     return dataset
 
 
-def grow_dataset(dataset: xr.Dataset):
+def grow_dataset(dataset: xr.Dataset) -> xr.Dataset:
     """
     Resizes the dataset by doubling the current length of all arrays.
 
@@ -270,23 +284,28 @@ def grow_dataset(dataset: xr.Dataset):
         The resized dataset
     """
     darrs = []
-    for col in dataset:
-        data = dataset[col].values
+
+    # coords will also be grown
+    for vname in dataset.variables.keys():
+        data = dataset[vname].values
         darrs.append(
             xr.DataArray(
-                name=dataset[col].name,
+                name=dataset[vname].name,
                 data=np.pad(data, (0, len(data)), "constant", constant_values=np.nan),
-                attrs=dataset[col].attrs,
+                attrs=dataset[vname].attrs,
             )
         )
+    coords = tuple(dataset.coords.keys())
     dataset = dataset.drop_dims(["dim_0"])
-    new_data = xr.merge(darrs)
-    return dataset.merge(new_data)
+    dataset = dataset.merge(xr.merge(darrs))
+    dataset = dataset.set_coords(coords)
+    return dataset
 
 
-def trim_dataset(dataset: xr.Dataset):
+def trim_dataset(dataset: xr.Dataset) -> xr.Dataset:
     """
-    Trim NaNs from a dataset, useful in the case of a dynamically resized dataset (eg. adaptive loops).
+    Trim NaNs from a dataset, useful in the case of a dynamically
+    resized dataset (e.g. adaptive loops).
 
     Parameters
     ----------
@@ -297,20 +316,97 @@ def trim_dataset(dataset: xr.Dataset):
     :class:`xarray.Dataset`
         The dataset, trimmed and resized if necessary or unchanged.
     """
+    coords = tuple(dataset.coords.keys())
     for i, val in enumerate(reversed(dataset["y0"].values)):
         if not np.isnan(val):
             finish_idx = len(dataset["y0"].values) - i
             darrs = []
-            for col in dataset:
-                data = dataset[col].values[:finish_idx]
+            # coords will also be trimmed
+            for vname in dataset.variables.keys():
+                data = dataset[vname].values[:finish_idx]
                 darrs.append(
                     xr.DataArray(
-                        name=dataset[col].name, data=data, attrs=dataset[col].attrs
+                        name=dataset[vname].name, data=data, attrs=dataset[vname].attrs
                     )
                 )
             dataset = dataset.drop_dims(["dim_0"])
-            new_data = xr.merge(darrs)
-            return dataset.merge(new_data)
+            dataset = dataset.merge(xr.merge(darrs))
+            dataset = dataset.set_coords(coords)
+            break
+
+    return dataset
+
+
+def to_gridded_dataset(
+    quantify_dataset: xr.Dataset,
+    dimension: str = "dim_0",
+    coords_names: Iterable = None,
+):
+    """
+    Converts a flattened (a.k.a. "stacked") dataset as the one generated by the :func:`~initialize_dataset`
+    to a dataset in which the measured values are mapped onto a grid in the `xarray` format.
+
+    This will be meaningful only if the data itself corresponds to a gridded measurement.
+
+    .. note:: Each individual :code:`(x0[i], x1[i], x2[i], ...)` setpoint must be unique.
+
+    Conversions applied:
+
+    - The names :code:`"x0", "x1", ...` will correspond to the names of the Dimensions.
+    - The unique values for each of the :code:`x0, x1, ...` Variables are converted to Coordinates.
+    - The :code:`y0, y1, ...` Variables are reshaped into a (multi-)dimensional grid and associated to the Coordinates.
+
+    .. seealso:: :meth:`~quantify.measurement.MeasurementControl.setpoints_grid`
+
+    Parameters
+    ----------
+    quantify_dataset
+        input dataset in the format generated by the :class:`~initialize_dataset`
+    dimension
+        the flattened xarray Dimension
+    coords_names
+        optionally specify explicitly which Variables correspond to orthogonal
+        coordinates, e.g. datasets holds values for :code:`("x0", "x1")` but only "x0"
+        is independent: :code:`to_gridded_dataset(dset, coords_names=["x0"])`
+
+    Returns
+    -------
+    :class:`xarray.Dataset`
+        the new dataset
+
+
+    .. include:: ./docstring_examples/quantify.data.handling.to_gridded_dataset.rst.txt
+    """
+
+    if coords_names is None:
+        # for compatibility with older datasets we use `variables` instead of `coords`
+        coords_names = sorted(
+            v for v in quantify_dataset.variables.keys() if v.startswith("x")
+        )
+    # Because xarray in general creates new objects and
+    # due to https://github.com/pydata/xarray/issues/2245
+    # the attributes need to be saved and restored in the new object
+    attrs_coords = tuple(quantify_dataset[name].attrs for name in coords_names)
+    # Convert "xi" variables to Coordinates
+    dataset = quantify_dataset.set_coords(coords_names)
+
+    # Convert to a gridded xarray dataset format
+
+    if len(coords_names) == 1:
+        # No unstacking needed just swap the dimension
+        for var in quantify_dataset.data_vars.keys():
+            if dimension in dataset[var].dims:
+                dataset = dataset.update(
+                    {var: dataset[var].swap_dims({dimension: coords_names[0]})}
+                )
+    else:
+        # Make the Dimension `dimension` a MultiIndex(x0, x1, ...)
+        dataset = dataset.set_index({dimension: coords_names})
+        # See also: http://xarray.pydata.org/en/stable/reshaping.html#stack-and-unstack
+        dataset = dataset.unstack(dim=dimension)
+    for name, attrs in zip(coords_names, attrs_coords):
+        dataset[name].attrs = attrs
+
     return dataset
 
 
@@ -339,14 +435,16 @@ def get_latest_tuid(contains: str = "") -> TUID:
     FileNotFoundError
         No data found
     """
-    return get_tuids_containing(contains, max_results=1)[0]
+    # `max_results=1, reverse=True` makes sure the tuid is found efficiently asap
+    return get_tuids_containing(contains, max_results=1, reverse=True)[0]
 
 
 def get_tuids_containing(
     contains: str,
-    t_start: datetime.date = None,
-    t_stop: datetime.date = None,
+    t_start: Union[datetime.datetime, str] = None,
+    t_stop: Union[datetime.datetime, str] = None,
     max_results: int = sys.maxsize,
+    reverse: bool = False,
 ) -> list:
     """
     Returns a list of tuids containing a specific label.
@@ -358,14 +456,20 @@ def get_tuids_containing(
 
     Parameters
     ----------
-    contains : str
+    contains
         a string contained in the experiment name.
-    t_start : datetime.date
-        date to search from, inclusive.
-    t_stop : datetime.date
-        date to search until, exclusive.
-    max_results : int
+    t_start
+        datetime to search from, inclusive. If a string is specified, it will be
+        converted to a datetime object using :func:`dateutil.parser.parse`.
+        If no value is specified, will use the year 1 as a reference t_start.
+    t_stop
+        datetime to search until, exclusive. If a string is specified, it will be
+        converted to a datetime object using :func:`dateutil.parser.parse`.
+        If no value is specified, will use the current time as a reference t_stop.
+    max_results
         maximum number of results to return. Defaults to unlimited.
+    reverse
+        if False, sorts tuids chronologically, if True sorts by most recent.
     Returns
     -------
     list
@@ -376,10 +480,24 @@ def get_tuids_containing(
         No data found
     """
     datadir = get_datadir()
+    if isinstance(t_start, str):
+        t_start = parse(t_start)
+    elif t_start is None:
+        t_start = datetime.datetime(1, 1, 1)
+    if isinstance(t_stop, str):
+        t_stop = parse(t_stop)
+    elif t_stop is None:
+        t_stop = datetime.datetime.now()
 
     # date range filters, define here to make the next line more readable
-    lower_bound = lambda x: x >= t_start if t_start else True  # noqa: E731
-    upper_bound = lambda x: x < t_stop if t_stop else True  # noqa: E731
+    d_start = t_start.strftime("%Y%m%d")
+    d_stop = t_stop.strftime("%Y%m%d")
+
+    def lower_bound(dir_name):
+        return dir_name >= d_start if d_start else True  # noqa: E731
+
+    def upper_bound(dir_name):
+        return dir_name <= d_stop if d_stop else True  # noqa: E731
 
     daydirs = list(
         filter(
@@ -389,41 +507,43 @@ def get_tuids_containing(
             os.listdir(datadir),
         )
     )
-    daydirs.sort(reverse=True)
+    daydirs.sort(reverse=reverse)
     if len(daydirs) == 0:
-        err_msg = 'There are no valid day directories in the data folder "{}"'.format(
-            datadir
-        )
+        err_msg = f"There are no valid day directories in the data folder '{datadir}'"
         if t_start or t_stop:
-            err_msg += ", for the range {}-{}".format(t_start or "", t_stop or "")
+            err_msg += f", for the range {t_start or ''} to {t_stop or ''}"
         raise FileNotFoundError(err_msg)
 
     tuids = []
-    for dd in daydirs:
+    for daydir in daydirs:
         expdirs = list(
             filter(
-                lambda x: (len(x) > 25 and TUID.is_valid(x[:26]) and contains in x),
-                os.listdir(os.path.join(datadir, dd)),
+                lambda x: (
+                    len(x) > 25
+                    and TUID.is_valid(x[:26])  # tuid is valid
+                    and (contains in x)  # label is part of exp_name
+                    and (t_start <= parse(x[:15]))  # tuid is after t_start
+                    and (parse(x[:15]) < t_stop)  # tuid is before t_stop
+                ),
+                os.listdir(os.path.join(datadir, daydir)),
             )
         )
-        expdirs.sort(reverse=True)
+        expdirs.sort(reverse=reverse)
         for expname in expdirs:
             # Check for inconsistent folder structure for datasets portability
-            if dd != expname[:8]:
+            if daydir != expname[:8]:
                 raise FileNotFoundError(
-                    'Experiment container "{}" is in wrong day directory "{}" '.format(
-                        expname, dd
-                    )
+                    f"Experiment container '{expname}' is in wrong day directory '{daydir}'"
                 )
             tuids.append(TUID(expname[:26]))
             if len(tuids) == max_results:
                 return tuids
     if len(tuids) == 0:
-        raise FileNotFoundError('No experiment found containing "{}"'.format(contains))
+        raise FileNotFoundError(f"No experiment found containing '{contains}'")
     return tuids
 
 
-def snapshot(update: bool = False, clean: bool = True) -> dict:
+def snapshot(update: bool = False, clean: bool = True) -> OrderedDict:
     """
     State of all instruments setup as a JSON-compatible dictionary (everything that the custom JSON encoder class
     :class:`qcodes.utils.helpers.NumpyJSONEncoder` supports).
@@ -436,10 +556,12 @@ def snapshot(update: bool = False, clean: bool = True) -> dict:
         if True, removes certain keys from the snapshot to create a more readable and compact snapshot.
     """
 
-    snap = {
-        "instruments": {},
-        "parameters": {},
-    }
+    snap = OrderedDict(
+        {
+            "instruments": {},
+            "parameters": {},
+        }
+    )
     for ins_name, ins_ref in Instrument._all_instruments.items():
         snap["instruments"][ins_name] = ins_ref().snapshot(update=update)
 
@@ -463,6 +585,10 @@ def snapshot(update: bool = False, clean: bool = True) -> dict:
 
 
 # ######################################################################
+# Private utilities
+# ######################################################################
+
+
 def _xi_and_yi_match(dsets: Iterable) -> bool:
     """
     Checks if all xi and yi data variables in `dsets` match:
@@ -487,17 +613,32 @@ def _vars_match(dsets: Iterable, var_type="x") -> bool:
     def get_xi_attrs(dset):
         # Hash is used in order to ensure everything matches:
         # name, long_name, unit, number of xi
-        return tuple(
-            dset[xi].attrs for xi in sorted(get_keys_containing(dset, var_type))
-        )
+        return tuple(dset[xi].attrs for xi in _get_parnames(dset, var_type))
 
-    it = map(get_xi_attrs, dsets)
+    iterator = map(get_xi_attrs, dsets)
     # We can compare to the first one always
-    tup0 = next(it, None)
+    tup0 = next(iterator, None)
 
-    for tup in it:
+    for tup in iterator:
         if tup != tup0:
             return False
 
     # Also returns true if the dsets is empty
     return True
+
+
+def _get_parnames(dset, par_type):
+    attr = "coords" if par_type == "x" else "data_vars"
+    return sorted(key for key in getattr(dset, attr).keys() if key.startswith(par_type))
+
+
+def _is_batched(obj) -> bool:
+    """
+    N.B. This function cannot be imported from quantify.measurement.type due to
+    some circular dependencies that it would create in the quantify.measurement.__init__
+
+    Returns
+    -------
+        the `.batched` attribute of the settable/gettable `obj`, `False` if not present.
+    """
+    return getattr(obj, "batched", False)